Creating new methods and binding classes together

+++ Library_Project/Book.java
@@ -1,12 +1,12 @@
 package Library_Project;
 
 public class Book {
-    public int serialNumber;
+    public long serialNumber;
     public String bookName;
     public String author;
     public int quantity;
     
-    public Book(int serialNumber, String bookName, String author, int quantity) {
+    public Book(long serialNumber, String bookName, String author, int quantity) {
         this.serialNumber = serialNumber;
         this.bookName = bookName;
         this.author = author;
